# ARROW-10173: [Rust][DataFusion] Implement support for direct comparison to scalar values