--- a/rust/datafusion/src/physical_plan/filter.rs
+++ b/rust/datafusion/src/physical_plan/filter.rs
@@ -128,7 +128,7 @@ fn batch_filter(
 ) -> ArrowResult<RecordBatch> {
     predicate
         .evaluate(&batch)
-        .map(|v| v.to_array(batch))
+        .map(|v| v.into_array(batch))
         .map_err(DataFusionError::into_arrow_external_error)
         .and_then(|array| {
             array