… rust/datafusion/Cargo.toml
… rust/datafusion/src/physical_plan/functions.rs
… rust/datafusion/src/physical_plan/hash_aggregate.rs
… rust/datafusion/src/physical_plan/projection.rs
… rust/datafusion/src/scalar.rs
… rust/datafusion/tests/sql.rs