@yordan-pavlov 2020-11-15
> I tried to run this code against our simple TPCH Q1 implementation (which is dominated by evaluating expressions against constants), and sadly it hit an error

@alamb it might be necessary to fallback to generating an array where the scalar value is repeated, for some operations that do not have a version which accepts a scalar argument