@alamb 2020-11-16
> @alamb it might be necessary to fallback to generating an array where the scalar value is repeated, for some operations that do not have a version which accepts a scalar argument

@yordan-pavlov  I agree with this plan. While long term I would like to see the ability to evaluate *every* function with constant arguments, that is definitely beyond the scope of this PR, and so keeping the fall back to constant arrays is a good idea. 

Given how important this functionality is, I suggest we focus on getting this PR into master as soon as possible and then iterating on additional functionality like additional function support, converting to canonical `expr _op_ colref` form, constant folding, hoisting ColumValue into arrow, etc. There is so much great work to do!