@yordan-pavlov 2020-11-16
@alamb I have now implemented falling back to scalar arrays for operations where scalar arguments are currently not supported; this should now work for operators such as "-" where previously an error was returned ("Scalar values on left side of operator - are not supported"); I have also added an extra test for this; 
could you try running the TPCH test again?