@alamb 2020-11-16
@yordan-pavlov  When I ran the benchmark locally again on my laptop:

```
cargo run --release --bin tpch -- --iterations 3 --path /Users/alamb/Software/tpch_data/SF10-parquet-64 --format parquet --query 1 --batch-size 4096
```

benchmarks on this branch (fd77fd6a4ee4705f8d25436c6a8dee3beb834941)
```
Running benchmarks with the following options: TpchOpt { query: 1, debug: false, iterations: 3, concurrency: 2, batch_size: 4096, path: "/Users/alamb/Software/tpch_data/SF10-parquet-64", file_format: "parquet", mem_table: false }
Query 1 iteration 0 took 5371 ms
Query 1 iteration 1 took 6121 ms
Query 1 iteration 2 took 6160 ms
alamb@ip-192-168-0-133 benchmarks %
```

benchmark on master oe5fce7f6a70c370c758b133f025444c98cdd305d

```
Running benchmarks with the following options: TpchOpt { query: 1, debug: false, iterations: 3, concurrency: 2, batch_size: 4096, path: "/Users/alamb/Software/tpch_data/SF10-parquet-64", file_format: "parquet", mem_table: false }
Query 1 iteration 0 took 7640 ms
Query 1 iteration 1 took 7808 ms
Query 1 iteration 2 took 7765 ms
```

So it seems to me the the performance improved a bit. I am second guessing my performance setup here - it is with a laptop and I do wonder if my CPU is being throttled.

Regardless, I think this is good enough results!