@yordan-pavlov 2020-11-17
@alamb thanks for running that test again; even if comparisons against scalar values are significantly faster, there is much more to the tests, such as loading data, etc.; I will be looking at that next. Overall though, good result as you said.